requirements.txt
colorlog==6.7.0
homeassistant==2024.2.0
pip>=21.0,<23.2
ruff==0.0.292
geopy>=2.1
python-dateutil>=2.8.0